src/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.18
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
GDAL==3.0.4
geojson==3.2.0
inflection==0.5.1
numpy==1.24.4
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1
